screen/requirements.txt
argparse
pillow
spidev
rpi.gpio
requests